requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.1
Adafruit-SSD1306==1.6.1
olefile==0.44
Pillow==4.2.1
RPi.GPIO==0.6.3
spidev==3.2
